pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neuralchild"
version = "0.1.0"
description = "A psychological brain simulation"
readme = "README.md"
authors = [{ name = "NeuralChild Team" }]
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "torch>=2.0.0",
    "pydantic>=2.0.0",
    "requests>=2.25.0",
    "pyyaml>=6.0",
    "numpy>=1.20.0",
]
requires-python = ">=3.8"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "mypy>=0.9.0",
]

[project.scripts]
neuralchild = "neuralchild.cli:main"

[tool.setuptools]
packages = ["neuralchild"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88